Add module for clicking users and channels
