+++ client/js/user-events/index.js
@@ -1,2 +1,4 @@
 "use strict";
 require("./message");
+require("./nick");
+require("./user-chan-click");
